what is wrong with that code?
The 7th floor of the array (the building) should hold the most significant bit while floor 0 should
have the least significant bit.
d. If the number was negative, you should convert the bit pattern to its complement. For that,
go through all the “floors” and invert them. If the floor had 1, then make it 0. If it has 0, then have
1. The next step should be adding 0000 0001 (1 in decimal) to it.
e. Step d means you need to write code that adds two binary numbers (stored in an array
(list)).
f. Once you are done with Step e, you are done with the complement operation. As your
final step, print the 1’s and 0’s on each floor starting with the most significant bit towards the least
significant bit.